+++ src/test/java/com/wootube/ioi/domain/model/UserTest.java
@@ -1,4 +1,4 @@
-package com.wootube.ioi.domain;
+package com.wootube.ioi.domain.model;
 
 import org.assertj.core.api.AbstractBooleanAssert;
 import org.assertj.core.api.AssertionsForClassTypes;
@@ -6,13 +6,12 @@ import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 
-import java.util.Set;
-import java.util.stream.Collectors;
-
 import javax.validation.ConstraintViolation;
 import javax.validation.Validation;
 import javax.validation.Validator;
 import javax.validation.ValidatorFactory;
+import java.util.Set;
+import java.util.stream.Collectors;
 
 public class UserTest {
     public static final String VALID_NAME = "루피";
